updated meeting
