+++ packages/teams-js/src/public/index.ts
@@ -66,7 +66,7 @@ export * as menus from './menus';
 export * as media from './media';
 export { secondaryBrowser } from './secondaryBrowser';
 export * as location from './location';
-export { meeting } from './meeting';
+export * as meeting from './meeting/meeting';
 export * as monetization from './monetization';
 export * as calendar from './calendar';
 export * as mail from './mail';
